+++ doc/changelog.rst
@@ -44,8 +44,9 @@ General
    point, as usual. If ``mjUSESINGLE`` is defined, MuJoCo will use single-precision floating point. See :ref:`mjtNum`.
 
    Relatedly, fixed various type errors that prevented building with single-precision.
-10. Quaternions in ``mjData->qpos`` are no longer normalized in-place by :ref:`mj_kinematics`. Instead they are
-    normalized when they are used. After the first step, quaternions will be normalized.
+10. Quaternions in ``mjData->qpos`` and ``mjData->mocap_quat`` are no longer normalized in-place by
+    :ref:`mj_kinematics`. Instead they are normalized when they are used. After the first step, quaternions in
+    ``mjData->qpos`` will be normalized.
 
 MJX
 ~~~